producer/requirements.txt
kafka-python==2.0.2
pymongo==3.11.0
